chore: extract-module ranks
